# doc/api.yml
openapi: 3.0.3
info:
  version: 'latest'
  title: HEI Admin API
  description: |
    _Programmatically connect to a computer programming [school](https://hei.school)._
    After [joining us](mailto:contact@mail.hei.school),
    you can get an identification token from
    our [dev](https://dev-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=5s8cg50doahmu855rlc8fr6qmp&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-dev.hei.school%2Fwhoami)
    or [prod](https://prod-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=i8bg538jpfu6mqmqb61m26trd&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-prod.hei.school%2Fwhoami)
    authentication service.
    Then, start playing with our system!

    The implementation of our API is [publicly disclosed](https://github.com/hei-school/hei-admin-api).
    You are welcome to try and compromise it. Happy hacking!
servers:
  - url: 'https://api-dev.hei.school'
  - url: 'https://api-prod.hei.school'
security:
  - BearerAuth: [ ]
paths:
...
  '/courses/{course_id}':
    get:
      tags:
        - Teaching
      summary: Get course by identifier
      parameters:
        - $ref: '#/components/parameters/PathCourseId'
      operationId: getCourseById
      responses:
        '200':
          description: The identified course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  '/courses/{course_id}/students/grades':
    get:
      tags:
        - Teaching
      summary: Get students final grades (all exams combined) for a specific course
      parameters:
        - $ref: '#/components/parameters/PathCourseId'
        - $ref: '#/components/parameters/GroupRef'
      operationId: getStudentsFinalGradesByCourseId
      responses:
        '200':
          description: Student list with their final grade for this course
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentGrade'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
  /courses:
    get:
      tags:
        - Teaching
      summary: Get all courses
      operationId: getCourses
      parameters:
        - name: code
          in: query
          description: "Filter courses by code, case is ignored"
          schema:
            type: string
        - name: name
          in: query
          description: "Filter courses by name, case is ignored"
          schema:
            type: string
        - name: credits
          in: query
          description: "Filter courses by credits"
          schema:
            type: integer
        - name: teacher_first_name
          in: query
          description: "Filter courses by teacher firstname, case is ignored"
          schema:
            type: string
        - name: teacher_last_name
          in: query
          description: "Filter courses by teacher lastname, case is ignored"
          schema:
            type: string
        - name: credits_order
          in: query
          schema:
            $ref: '#/components/schemas/CourseDirection'
        - name: code_order
          in: query
          schema:
            $ref: '#/components/schemas/CourseDirection'
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
          description: Set value to 1 by default if null is provided
          required: false
        - name: page_size
          in: query
          description: Set value to 15 by default if null is provided
          schema:
            $ref: '#/components/schemas/PageSize'
      responses:
        '200':
          description: List of all courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Teaching
      summary: Create new courses or update existing courses
      description: Update courses when `id` are provided, create them otherwise.
      operationId: CreateOrUpdateCourses
      requestBody:
        description: Courses to update
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Course'
      responses:
        '200':
          description: Course created or modified
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    GroupRef:
      name: group_ref
      in: query
      example: G1
      required: false
      description: Filter by group ref
      schema:
        type: string
...
    PathCourseId:
      name: course_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    UserIdentifier:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
          example: STD000001
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        nic:
          type: string
...
    CrupdateUser:
      allOf:
        - $ref: '#/components/schemas/UserIdentifier'
        - type: object
          properties:
            sex:
              $ref: '#/components/schemas/Sex'
            birth_date:
              type: string
              format: date
            birth_place:
              type: string
            address:
              type: string
            phone:
              type: string
            entrance_datetime:
              type: string
              format: date-time
            coordinates:
              $ref: '#/components/schemas/Coordinates'
            high_school_origin:
              type: string
            status:
              $ref: '#/components/schemas/EnableStatus'
    WorkStudyStatus:
      type: string
      enum:
        - HAVE_BEEN_WORKING
        - WILL_BE_WORKING
        - WORKING
        - NOT_WORKING
    SpecializationField:
      type: string
      enum:
        - COMMON_CORE
        - EL
        - TN
    Sex:
      type: string
      enum:
        - M
        - F
    EnableStatus:
      type: string
      enum:
        - ENABLED
        - DISABLED
        - SUSPENDED
...
    Student:
      allOf:
        - $ref: '#/components/schemas/CrupdateUser'
      properties:
        specialization_field:
          $ref: '#/components/schemas/SpecializationField'
        professional_experience:
          $ref: '#/components/schemas/ProfessionalExperienceFileTypeEnum'
        work_study_status:
          $ref: '#/components/schemas/WorkStudyStatus'
        commitment_begin_date:
          type: string
          format: date-time
        commitment_end_date:
          type: string
          format: date-time
        profile_picture:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        is_repeating_year:
          type: boolean
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: The group number one
        ref:
          type: string
          example: G1
        creation_datetime:
          type: string
          format: date-time
        size:
          type: integer
        attributed_color:
          type: string
...
    CourseTemplate:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
          example: PROG1
        name:
          type: string
          example: Algorithmics
        credits:
          type: integer
        total_hours:
          type: integer
    Course:
      allOf:
        - $ref: '#/components/schemas/CourseTemplate'
...
    Grade:
      type: object
      required:
        - score
      properties:
        id:
          type: string
        score:
          type: number
          format: double
          example: 12.5
        created_at:
          type: string
          format: date-time
        update_date:
          type: string
          format: date-time
...
    StudentGrade:
      properties:
        grade:
          $ref: '#/components/schemas/Grade'
        student:
          $ref: '#/components/schemas/Student'
...
    ProfessionalExperienceFileTypeEnum:
      type: string
      enum:
        - WORKER_STUDENT
        - BUSINESS_OWNER
        - INTERN_STUDENT
        - INTERN_PROJECT
...
    Coordinates:
      type: object
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
      default: 15
...
    CourseDirection:
      type: string
      enum:
        - ASC
        - DESC
...
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error